Major amendments to arrivial page - '/'
Majory amendments to '/' page.

Added OpenLayers, and began to focus on Auchterarder.  Added search bar at top of '/' which as yet only has local date (not from store), and is not yet functional.

Create NavBar component and Map component, both of which are works in progress.

Minor changes made to styling - colours / layouts changed.