Added GPX routes on map
Added routing on map, from GPX files.  Rather rough solution to this so far.  A lot of work still required to refine, but working at present.  Need to prevent info appearing when route is clicked, and not start point.  Potentially need to rethink whether to display the routes, or to only make them viewable when the way point is clicked.